+++ docs/changelog.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### 2026-04-03
+- CLI update: subcommand must be specified
 - ChatCompletionParameters add repeat_penalty and repeat_last_n
-- generate add penalty repeat code
+- generate add penalty repeat
 
 ### 2026-04-02
 - refactor generate code